+++ rats-tls/Cargo.toml
@@ -8,3 +8,4 @@ edition = "2021"
 [dependencies]
 foreign-types = "0.5.0"
 tokio = { version = "1.19.2", features = ["full"] }
+tokio-util = { version = "0.7.3", features = ["io-util"] }
